# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for the final deliverable",
        "title": "2022 Wizeline's Go Bootcamp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "lroman@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Hello World",
                "tags": [
                    "Description"
                ],
                "summary": "Hello World",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/characters": {
            "get": {
                "description": "Read Characters from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Character List DB",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Retrieve a Character by id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/charactersApi": {
            "get": {
                "description": "Read Characters from the Rick \u0026 Morty's API and wirtes the response into a CSV File",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Character List API",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/readCsv": {
            "get": {
                "description": "Read Characters from a Csv File",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Character List Csv",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/readCsvConcurrently": {
            "get": {
                "description": "Read Characters from a Csv File Concurrently",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Character List Csv",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Retrieve 'odd' or 'even' ids",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Quantity of items to retrieve",
                        "name": "items",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Quantity of items that each worker will read concurrently",
                        "name": "items_per_worker",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controller.HomeResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}